feature: Share stack-based inlining between TypeResolver and the new Typer
First step of migrating to the new Typer (#1764, #392). To avoid
maintaining two divergent inlining implementations, the cycle/depth
guarded function and partial-query inlining added in #1765 is extracted
from TypeResolver into a shared FunctionInliner, and the new Typer now
applies it inside its traversal:

- PartialQueryApply nodes are inlined before typing (plan-level rewrite)
- FunctionApply nodes and no-arg member references are inlined after
  typing, in separate passes so a member call's base DotRef is not
  visited before its enclosing FunctionApply
- Inlined bodies are re-typed so tpe fields stay consistent

A new TyperParityCheck gate compiles every spec/basic file with both
pipelines and compares generated SQL, pinning the current parity level
(44/54 standalone-compilable specs identical, up from 39 before this
change; remaining gaps are grouping-key indexes, aggregation-function
inlining, native expressions, and TableRef resolution). Improvements
pass automatically; regressions fail the build.

TypeResolver behavior is unchanged (pure extraction).
